fix gtid reset (#115)
- fix bug on gtid start position

when the spec has start position, we actually didn't respect the spec's start position; and the test was inconsistent with the intension.

- fix bug on int primary key scan.

When primary key is int, the scanned value was converted to string, and broke the comparison using select a > b

+++ pkg/mysql_test/test.go
@@ -39,7 +39,7 @@ const (
 
 var setupSqls = []string{
 	fmt.Sprintf(`CREATE TABLE IF NOT EXISTS %s (
-  id int(11) unsigned NOT NULL,
+  id bigint(20) NOT NULL,
   name varchar(256) DEFAULT NULL,
   email varchar(30) COLLATE utf8mb4_bin NOT NULL DEFAULT 'default_email',
   ts TIMESTAMP,
